feat/BE: 예외 controller advice 공통 오류 페이지 반환하도록 수정 (#364)
